DistributedLog: Allow Instantiating DistributedLogManagerFactory to manage all the log streams in the app Motivation is use cases where a single process has to manage access to a large number of log containers at the same time.
Instantiate a Factory which provides interfaces to
a) Create managers for individual containers
b) Enumerate all containers within a namespace
c) Manages connections to ZooKeeper and BookKeeper that are shared by all the containers. We only have one connection per factory at the moment. For more high performance scenarios we may need to have a pool of connections.

RB_ID=169115