DistributedLog: Optimize resume Motivated by high CPU utilization seen in non blocking reads over very low volume streams
Positioning recreates the LedgerInputStream and the Reader. This is only necessary when a new ledger handle is opened or when an inprogress segment is completed and we must now read from the closed ledger. In all other cases we can simply re-use the already initialized LedgerInputStream and Reader.

RB_ID=261831