DistributedLog: Handle missing ZK notification Detect the case where the read ahead view of the log segment metadata is stale when a zookeeper notification has been missed and trigger reinitialization of metadata
The case we detect is when the current ledger handle is marked as closed and the corresponding metadata is still in progress. We detect this if the state persists for idle reader warn threshold.

Since we always close the ledger first and then update the log segment metadata its fine if we see this state, however it should be short lived. Thats why the logic to wait for the idle warn threshold. If we forced metadata to be reinitialized as soon as we detected a closed ledger with in progress metadata, we may issue unnecessary ZK reads.

RB_ID=542680